endpoint price

+++ migrations/20240503093316-create-game.js
@@ -13,7 +13,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL
       },
       players: {
         type: Sequelize.INTEGER
